sample_8/requirements.txt
annotated-types==0.6.0
pydantic==2.5.2
pydantic_core==2.14.5
PyYAML==6.0.1
typing_extensions==4.8.0
